class9/QuoteForge/pyproject.toml
[project]
name = "quoteforge"
version = "0.1.0"
description = "A beautiful quote poster generator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.45.1",
    "Pillow>=11.2.1",
    "numpy>=2.2.6",
    "requests>=2.31.0",
    "python-unsplash>=1.1.0",
    "pyjwt>=2.8.0",
]
